change data type of price to float

+++ database/migrations/2024_11_14_090052_create_fixed_costs_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description')->nullable();
-            $table->bigInteger('amount');
+            $table->float('amount');
             $table->string('frequency')->default('monthly');
             $this->getDefaultTimestamps($table);
         });
